pyproject.toml
[tool.poetry]
name = "mediasoft-httpservice"
version = "0.1.0"
description = ""
authors = ["ajib6ept <asalbert@narod.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"
coverage = "^6.5.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
